refactor: Post,CommentApiController Mapper 패키지 분리
레이어에 종속적이지 않고 각 도메인 영역에 `mapper` 영역을 통해 관리

Controller가 다른 Service를 바라보는 경우 양방향 의존성이 생김
이를 방지하는 방법으로는 특정 레이어에 맞는 mapper로 구현하는 방법이 있지만, mapper가 특정 레이어에 맞게 구현하면 많은 책임을 지게 됨
`mapper` 패키지로 분리하고 추후 리팩토링
