+++ src/main/java/com/nubble/backend/comment/controller/CommentApiController.java
@@ -4,6 +4,8 @@ import com.nubble.backend.comment.controller.CommentRequest.GuestCommentCreateRe
 import com.nubble.backend.comment.controller.CommentRequest.GuestCommentDeleteRequest;
 import com.nubble.backend.comment.controller.CommentResponse.CommentCreateResponse;
 import com.nubble.backend.comment.controller.CommentResponse.CommentFindResponses;
+import com.nubble.backend.comment.mapper.CommentCommandMapper;
+import com.nubble.backend.comment.mapper.CommentResponseMapper;
 import com.nubble.backend.comment.service.CommentCommand.CommentCreateCommand;
 import com.nubble.backend.comment.service.CommentCommand.CommentDeleteCommand;
 import com.nubble.backend.comment.service.CommentInfo;
