+++ src/main/java/com/nubble/backend/post/mapper/PostResponseMapper.java
@@ -1,4 +1,4 @@
-package com.nubble.backend.post.controller.mapper;
+package com.nubble.backend.post.mapper;
 
 import com.nubble.backend.post.controller.PostResponse.MemberCommentCreateResponse;
 import com.nubble.backend.post.controller.PostResponse.PostCreateResponse;
